Ongoing damage tokens
#27 Added poison, fire, bleed, and acid effects that deal damage based on their duration.

Moved damage handling to its own script, and branched the damage workflow for units so that there aren't chat notifications for "Unconscious" when a unit is broken. This could cause problems if damage handling is radically changed in the future, but it should only break for unit damage, not for npc/pc damage, that stays exactly the same.