CT Quality of Life
The GM can now drag/drop a CT unit's link icon, and if it is dropped on a non-unit's CT entry, the unit will be 'assigned' to that commander. This does the following: sets the unit's commander to the drop target, sets the unit's initiative to just after the commanders, and sets the unit's faction to match the commander.

In the future this can also do a deep share if the faction is friendly to players.