Is Unit Exposed Algorithm
Added algorithm to determine if units are exposed. Algorithm seems to be performant, though it could get dicey with tons of units.

I also haven't tested it with a map that is oriented differently.

Todo: Make the markers use string:match to get just the token name (not the path) instead of literal string matches.