Tentative fix for map markers on clients
Tentative fix for fortification tokens not being public and thus not working when clients attempt to roll attacks targeting units on top of said fortifications.

This fix of course requires the GM to load the fortification window at least once to run onInit(), but that should always be the case since fortifications begin as empty.