Add commander link to unit sheet
Unit reference sheet now contains a link to the commander. This link is set when units are add to the combat tracker and their commander is already on the CT, and when units are assigned to a commander in the CT.

When the link is set, the commander name field is forced to read-only, but will update itself when the commander's name is updated.

If the commander CT entry is deleted, the link is removed and the name is returned to being editable.