Update manager_combat_kw.lua
Replace all of the addPC madness with a simple function override. We're not replacing the addPC functionality, just adding to it so there's no reason to set a custom addPC function. This should be better at working with other extensions as well. Might need to look at how the return value is handled in the future.